requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.4
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
idna==3.10
psycopg==3.2.3
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.2.3
